[RFC PATCH] phc2sys: Add multi-domain sync
This patch adds support for synchronization between clocks controlled by
different ptp4l instances. phc2sys can now communicate with multiple
ptp4l instances (specified with multiple -z options). Clocks from one
instance are considered a domain. If the domain doesn't have a source
clock (port in the slave state), the clocks in the domain can be
synchronized to a source clock from a different domain. The real-time
clock is a separate domain, which can be a source for other domains if
the -r option is specified twice.

This allows the system clock or a PHC to be switched to a backup source
on some (e.g. network) failures. The selection of the source could
include a configurable priority or compare specific PTP attributes.
Currently, it's the first synchronized source in the order of specified
sockets.

TODO:
- fix non-automatic modes
- extend configuration to support domain-specific settings
- split into multiple patches
- ...

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
.